pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-maestro"
version = "0.1.0"
description = "Multi-agent task orchestration — delegate work from an orchestrator to specialized sub-agents via a file-based queue."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
]

[project.scripts]
maestro = "agent_maestro.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[tool.setuptools.packages.find]
include = ["agent_maestro*"]

[tool.setuptools.package-data]
agent_maestro = [
    "scaffold/**/*",
    "scaffold/**/*.yaml",
    "scaffold/**/*.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
